Bugfix: fix nullptr if project has empty active models entry

+++ backend/src/main/java/de/hskl/ki/services/ProjectService.java
@@ -87,7 +87,7 @@ public class ProjectService {
     private void deleteProjectFolder(Project project) {
         containerProxyService.stopContainer(project.getId());
 
-        if(!project.getActiveModels().isEmpty()) {
+        if(project.getActiveModels() != null && !project.getActiveModels().isEmpty()) {
             inferenceService.deactivateModel(project.getActiveModels());
             inferenceService.deleteModelFromTritonFolder(project.getActiveModels());
         }
